refactor: move tetrominos to constants file
